Split conftest.py to base_conftest.py and conftest.py. Adapted Docs. test.base_conftest is not a plugin. base_conftest.py includes the base to start testing vyper contracts while conftest.py includes all the other fixtures and functions needed to test our code.
